requirements.txt
python==3.12.9
dill==0.3.8
future==1.0.0
matplotlib==3.10.7
nltk==3.9.1
numpy==2.3.4
pandas==2.3.3
python_dateutil==2.9.0.post0
scikit_learn==1.7.2
seaborn==0.13.2
